Исправления безопасности, багов и оптимизация
Backend:
- orderController: авторизация на всех под-ресурсах (checkOrderAccess), удаление файлов с диска при удалении записей
- chatController: проверка доступа к orderId/clientId в getMessages
- documentController: валидация принадлежности orderId/clientId при создании
- uploadController: авторизация на удаление смет, очистка файлов с диска
- adminController: запрет удаления других администраторов
- authController: реализован endpoint changePassword
- token.js: предупреждение при отсутствии JWT_SECRET
- models/index.js: CASCADE правила, связь Design-User, Order.assignedTo-User
- CostEstimate: FLOAT -> DECIMAL(12,2) для extractedAmount

Frontend:
- auth.js: try-catch для JSON.parse localStorage
- app.js: очистка полей логина, защита от дублирования обработчиков, вызов destroy при навигации
- chat.js: очистка refreshTimer при уходе со страницы (memory leak fix)
- profile.js: реализована смена пароля через API

+++ backend/models/CostEstimate.js
@@ -20,7 +20,7 @@ const CostEstimate = sequelize.define('CostEstimate', {
     allowNull: false,
   },
   extractedAmount: {
-    type: DataTypes.FLOAT,
+    type: DataTypes.DECIMAL(12, 2),
     allowNull: true,
   },
   uploadedAt: {
